+++ src/components/Cart/index.ts
@@ -1,0 +1,2 @@
+export { default as Cart } from './Cart';
+export { default as CartItem } from './CartItem';
